fs:yaffs The return value of yaffsfs_CurrentTime is truncated
VELAPLATFO-7125

dependson:3499458

(cherry picked from commit 46481fda62764010e938790bda13675e1bb0f9a6)

+++ fs/yaffs/yaffs_osglue.c
@@ -56,7 +56,7 @@ void yaffsfs_Unlock(void)
   nxmutex_unlock(&g_yaffs_lock);
 }
 
-u32 yaffsfs_CurrentTime(void)
+u64 yaffsfs_CurrentTime(void)
 {
   return time(NULL);
 }
